Assignment3/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-QRcode==3.0.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Pillow==8.4.0
pycryptodome==3.11.0
qrcode==7.3.1
requests==2.26.0
urllib3==1.26.7
Werkzeug==2.0.2